rename rulename.
